requirements.txt
crewai
duckduckgo-search
langchain_google_genai
langchain-community
streamlit
python-dotenv